refactor: Answer서비스로 분리
